fix(pypi): Exclude known crawlers from download stats
PyPI download numbers are somewhat skewed by automated
crawlers that periodically download assets from PyPI.

Examples include the `bandersnatch` PyPI mirror client, which
accounts for a relatively high fraction of all downloads (~25%).

This commit adds a list of known crawlers to be excluded to the
BigQuery query in `PyPIMetrics.downloads()`.