refactor: Refactor BigQueryWriter to use merge query for data updates
Refactored the `BigQueryWriter` class in `db.py` to use a merge query for updating data in the destination table. This allows for more efficient and flexible updates, as well as handling the case when the destination table does not exist. The merge query is used to merge the data from a temporary table into the destination table, updating existing rows and inserting new rows as needed. The temporary table is then deleted after the merge is complete.