fix: Improve GH issues metrics calculation time
By disabling the detection of linked PRs for issues, we can greatly
reduce the time taken to calculate GH issues metrics.

This means that the `has_pull_request` field will be set to `pd.NA` for
all issues, as the API calls to check for linked PRs are expensive and
not always necessary for the current metrics.